Readme and setup files
Filled in more details on the readme.md file (as referenced throughout file).

Began setup files for Windows and Linux/Mac. Windows (setup.bat) is working correctly but has a strange output line. Mac has yet to be tested. As well, the setup files now prompt the user to input db connection settings. The code now reflects that dynamic input. The DB information is now stored in the .env file (which is created on setup.bat running). Gitignore was added to hide files like .env but that also isn't set up yet.