Implementing bcrypt and some port logic updates
bcrypt has been included for the password hashing. Reference to /register and /login. Users can now register and have the supporting logic to not create duplicates. As well, users can then login. This has been tested and confirmed with postman.

Ports! We have port 3000 now reserved for react and express recommended to be port 5000 (backend).

the setup.bat and sh files now reflect the above updates, preventing selection of port 3000, and including bcrypt in the installed dependencies list.