Full merge for this week March28th deliverable
Due to architecture changes, merging quickly became a nightmare through gitbash and git. Luckily, everyone's changes were isolated enough (in the type of work and in file location) that I could move the files over into my branch, then merging.

This is obviously not how it was intended to work but the correct solution would have been to change the architecture post submission last week, THEN clone it for this weeks dev for everyone to copy.

Lesson learned, code mostly corrected.

Controller code and module code is a bit messy at the moment. Ral and Marcus will agree on the files structure and make a more organized space