JWT to cookies! (I broke it)
This is an update for our deliverable. I say this, because I usually only update once functionality is achieved but couldn't quiet get it working this week in time.

After successfully working with bcrypt, JWT, and some react to backend end points, functionality was a success and /register, login, and the redirect to dashboard works! Right?

It did work, but it was storing the JWT in the Reacts storage (well it wasn't in the session or local storage after testing). This was a problem because it allows javascript access and then cross-site injections(?) .

After some further work, we have progressed to using and passing cookies! I have been working through the troubleshooting and working our way from backend to frontend. I'm currently having some issues passing it through the front end, but I just need a little more time to troubleshoot.

So for this week, the API endpoints have been confirmed to work (well, /register, /login, and /dashboard) but now only /register and /login work as /dashboard will deny you.