Updates post security update to frontend pages
/Login, /Register, and /Dashboard have all received updates post security update (cookies).

Pages have received cosmetic upgrades and now reroute after success in registration or login.

Login now doesn't include userId or type and will be later dealt with with the admin page.

Backend was update to respond to this, including adding new dependencies to deal with auto-incrementing the userId and changing some other minor details.