Frontend Pages organized and added
After researching topics from local municipalities, I have created an array of pages for the frontend that present a good scope for what we plan to fill in. Not all of these pages will be used, but all these include all of the pages to be used.

Welcome pages direct to their sub categories. Simple pages are many and complicated pages are few. Therefore, they are separated to avoid confusion.

Backend logic files will be next, but waiting on cleanup and fixing of controllers and models by other group members.