Manager page, some logic rework (model and controllers) and other stuff
This update was a mixed bag of one problem leading to another solution and or problem.

The login, registration, and Manager pages have all received react type styling.

The manager page now automatically pulls the users list with some choice data (not password).

The manager page also includes a delete button with accompanying backend logic to support the button.

On the side, I've left the messy model and controllers as they are and begun adding around it.